+++ 42/media/lua/client/ExcretionFunctions/BF_Defecation.lua
@@ -1,3 +1,5 @@
+require "BodilyFunctions"
+
 -- Function to update defecation-related values
 function BF.UpdateDefecationValues()
     local player = getPlayer()